% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inventory.R
\name{basic_inventory_request}
\alias{basic_inventory_request}
\title{Retrieve Basic Inventory Reports}
\usage{
basic_inventory_request(credentials,
                               request_type=c('Configuration', 'Overview',
                                   'Campaign', 'Campaign.Detail',
                                   'Page', 'Page.Detail',
                                   'Position', 'Position.Detail',
                                   'Section', 'Section.Detail',
                                   'Site', 'Site.Detail',
                                   'PageAtPosition', 'PageAtPosition.Detail',
                                   'SectionAtPosition', 'SectionAtPosition.Detail',
                                   'SiteAtPosition', 'SiteAtPosition.Detail'),
                               report_name,
                               keywords=NULL,
                               position=NULL,
                               id=NULL,
                               start_date=NULL,
                               end_date=NULL)
}
\arguments{
\item{credentials}{a character string as returned by \link{build_credentials}}

\item{report_type}{a character string in one of eighteen supported
inventory report types}

\item{report_name}{a character string of the report name. Please see the
dataset \link{inventory_reports} and filter to inventory_type=="basic" for
a complete list of available report_type and report_name combinations for this
function.}

\item{keywords}{a character string containing comma separated keywords value
for filtering inventory report.}

\item{position}{a character vector containing position names.
Multiple "position" elements be specified at once, hence a vector is accepted
instead of a string}

\item{id}{a character string This element is used for defining id
(Campaign, Site, Page, Section) for inquiring the inventory detail report. Only
necessary for detail report types (e.g. Position.Detail)}

\item{start_date}{a character string for inquiring inventory detail report.
This field is optional and must be in the yyyy-mm-dd format.}

\item{end_date}{a character string for inquiring the inventory detail report by
given schedule. This field is optional and must be in the yyyy-mm-dd format.}
}
\value{
A \code{data.frame} of inventory data in the format of the specified
report_type and report_name
}
\description{
This function returns a data.frame of inventory statistics for
18 different report types
}
\examples{
\dontrun{
my_credentials <- build_credentials('myaccount',
                                    'myusername',
                                    'mypassword')
config <- basic_inventory_request(credentials=my_credentials,
                                  report_type='Configuration',
                                  report_name='Configuration')
# note that forecast start and end dates must be greater than or equal to
Sys.Date() otherwise it will return a 0 row data.frame
overview <- basic_inventory_request(credentials=my_credentials,
                                    report_type='Overview',
                                    report_name='All Sites Forecast',
                                    start_date='2015-12-01',
                                    end_date='2015-12-31')
# leaving position argument NULL means all positions are returned
# specifying the position argument NULL means only those will be returned
page_pos_forecast <- basic_inventory_request(credentials=my_credentials,
                                             report_type='PageAtPosition.Detail',
                                             report_name='Detail Forecast',
                                             id='www.site.com/page@x01',
                                             start_date='2015-12-01',
                                             end_date='2015-12-31')
}
}
\concept{
api inventory report
}

