% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authentication.R
\name{build_credentials}
\alias{build_credentials}
\title{Build Credentials String}
\usage{
build_credentials(account, username, password)
}
\arguments{
\item{account}{a character string of your OAS account name}

\item{username}{a character string of your OAS username that has API enabled access permissions}

\item{password}{a character string of your password for the OAS username specified}
}
\value{
A \code{character}, formatted as XML, that should be supplied when making requests to the API
}
\description{
This function returns a formatted string that should be supplied
when making requests to the API to authenticate the user
}
\examples{
\dontrun{
my_credentials <- build_credentials(account='MYACCOUNTNAME', username='MYUSERNAME', password='MYPASSWORD')
}
}
\concept{
api login credentials
}

