% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{error_codes}
\alias{error_codes}
\title{Error Codes Table}
\format{a \code{data.frame} with 192 rows and 3 variables}
\usage{
error_codes
}
\description{
This data.frame contains error codes and explanations
}
\details{
\itemize{
  \item errorCode: a numeric code indicating an error type
  \item Type: a brief statement of the type of error
  \item Explanation: a reason why that type of error occured
}
}
\examples{
\dontrun{
data(error_codes)

head(error_codes)
#  errorCode                                       Type             Explanation
#1         0                                    Success Not normally presented.
#2       300 Authentication Failed: Incorrect Password.   Credentials incorrect
#3       302  Authentication Failed: Account Incorrect.   Credentials incorrect
}
}
\keyword{datasets}

