% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventory_reports}
\alias{inventory_reports}
\title{Inventory Reports Table}
\format{a \code{data.frame} with 619 rows and 3 variables}
\usage{
inventory_reports
}
\description{
This data.frame contains all of the inventory reports available via
the \link{*_inventory_request} functions. Simply provide the
same report_type and report_name as seen in this table
to retrive via those functions
}
\details{
\itemize{
  \item inventory_type: the type of inventory function being called
  (e.g. basic, search, geo, or zone)
  \item report_type: the type of object the report covers
  \item report_id: an OAS id needed internally to request the report table
  \item report_name: a human readable name of the report
}
}
\examples{
\dontrun{
data(inventory_reports)

head(inventory_reports)
#          report_type                      report_id                       report_name
#  1 Campaign Delivery Delivery.Campaign.Base.T100.03                 Executive Summary
#  2 Campaign Delivery Delivery.Campaign.Base.T102.06                  Campaign Summary
#  3 Campaign Delivery Delivery.Campaign.Base.T110.02              Campaign Information
#  4 Campaign Delivery Delivery.Campaign.Base.T120.02   Advertiser & Agency Information
#  5 Campaign Delivery Delivery.Campaign.Base.T150.01 Day of Month Delivery Information
#  6 Campaign Delivery Delivery.Campaign.Base.T152.01  Day of Week Delivery Information
}
}
\keyword{datasets}

