% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_result_parser}
\alias{list_result_parser}
\title{Parse List Recordsets with Pagination Support}
\usage{
list_result_parser(xmlBody, paginate, result_text, request_type, verbose = FALSE)
}
\arguments{
\item{xmlBody}{a character string that appears as XML of the original request
sent over to the API}

\item{paginate}{a \code{logical} indicating whether pagination is needed}

\item{result_text}{a character string that appears as XML as 
returned by an API call}

\item{request_type}{a character string in one of the supported 
object types for the API database list action}

\item{verbose}{a boolean indicating whether to print the request XML as a message}
}
\value{
A \code{data.frame} listing all objects of the specified type that 
also met the supplied search criteria
}
\description{
This function takes an initial API return message with parameters
on how to paginate and parse the message and subsequent calls if needed
}
\concept{
api list
}

