% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{oas_add}
\alias{oas_add}
\title{Add Items in API Database}
\usage{
oas_add(credentials, 
                    request_type=c('Advertiser', 'AdvertiserCategory', 'Affiliate',
                                   'Agency', 'CampaignGroup', 'CompanionPosition',
                                   'CompetitiveCategory', 'ConversionProcess', 
                                   'CreativeType',
                                   'Event', 'InsertionOrder', 'Notification',
                                   'Page', 'Product', 'RichMediaTemplate', 
                                   'SalesPerson', 'Section', 'Site',
                                   'SiteGroup', 'Transaction', 'Position',
                                   'Keyword', 'Keyname', 'Publisher', 
                                   'Campaign', 
                                   'CreativeTarget', 'Creative'),
                     add_data, verbose = FALSE)
}
\arguments{
\item{credentials}{a character string as returned by \link{oas_build_credentials}}

\item{request_type}{a character string in one of the supported 
object types for the API database list action}

\item{add_data}{an XML document specifying the data to add when creating the request_type}

\item{verbose}{a boolean indicating whether messages should be printed while making the request}
}
\value{
A \code{logical} being TRUE if the add was successful
}
\description{
This function adds items in Xaxis for Publishers.
}
\note{
We recommend using the \link{oas_list} function to first understand
what attributes are available on each object type in Xaxis for Publishers
before attempting to add existing objects, especially for complicated types, 
such as, Campaigns which have many configurable attributes.
}
\examples{
\dontrun{
my_credentials <- oas_build_credentials('myaccount', 
                                        'myusername', 
                                        'mypassword')
                                    
adver_add <- oas_add(credentials=my_credentials, 
                           request_type='Advertiser', 
                           update_data=list(addChildren(
                                            newXMLNode('Advertiser'), 
                                               list(newXMLNode('Id', 'MyAdvertiserId'), 
                                                    newXMLNode('ContactTitle', 'new Title')))))
                                            
}
}
\concept{
api add
}

