% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{oas_build_credentials}
\alias{oas_build_credentials}
\title{Build Credentials String}
\usage{
oas_build_credentials(account=getOption("roas.account"), 
                         username=getOption("roas.username"), 
                         password=getOption("roas.password"))
}
\arguments{
\item{account}{a character string of your OAS account name}

\item{username}{a character string of your OAS username that has API enabled access permissions}

\item{password}{a character string of your password for the OAS username specified}
}
\value{
A \code{character}, formatted as XML, that should be supplied when making requests to the API
}
\description{
This function returns a formatted string that should be supplied 
when making requests to the API to authenticate the user
}
\examples{
\dontrun{
my_credentials <- oas_build_credentials(account='myaccountname', 
                                    username='myusername', 
                                    password='mypassword')
}
}
\concept{
api login credentials
}

