% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{oas_copy}
\alias{oas_copy}
\title{Copy an Item in API Database}
\usage{
oas_copy(credentials, 
                    request_type='Campaign',
                     copy_attributes = NULL,
                     copy_data, verbose = FALSE)
}
\arguments{
\item{credentials}{a character string as returned by \link{oas_build_credentials}}

\item{request_type}{a character string in one of the supported 
object types for the API database list action}

\item{copy_attributes}{a named character vector of attributes 
to add to the Copy node. The only acceptable parameter is "recurring".}

\item{copy_data}{an XML document specifying the data to copy when creating the request_type}

\item{verbose}{a boolean indicating whether messages should be printed while making the request}
}
\value{
A \code{logical} being TRUE if the copy was successful
}
\description{
This function copies items in Xaxis for Publishers.
}
\note{
See the documentation about creating recurring copies
}
\examples{
\dontrun{
my_credentials <- oas_build_credentials('myaccount', 
                                        'myusername', 
                                        'mypassword')
                                    
campgn_copy <- oas_copy(credentials=credentials, 
                        request_type='Campaign', 
                        copy_data=list(newXMLNode('Id', 'oldCampaignId'), 
                                         newXMLNode('NewId', 'newCampaignId'), 
                                         newXMLNode('CopyCreatives', 'Y'), 
                                         newXMLNode('CopyNotifications', 'Y'), 
                                         newXMLNode('CopyScheduling', 'Y'), 
                                         newXMLNode('CopyTargeting', 'Y'),
                                         newXMLNode('CopyBilling', 'Y'), 
                                         newXMLNode('CopyPages', 'Y'), 
                                         newXMLNode('CopySiteTiers', 'Y'), 
                                         newXMLNode('CopyConversionProcesses', 'Y')))
}
}
\concept{
api copy
}

