% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{oas_read}
\alias{oas_read}
\title{Read/Retrieve all details for a requested object}
\usage{
oas_read(credentials, 
                    request_type=c('Advertiser', 'AdvertiserCategory', 
                                   'Affiliate', 'Agency', 'CampaignGroup', 
                                   'ConversionProcess', 'Zone', 'CreativeType',
                                   'Event', 'InsertionOrder', 'RichMediaTemplate', 
                                   'Product', 'SalesPerson', 
                                   'Site', 'SiteGroup', 'Publisher', 
                                   'Page', 'Section', 'Transaction',
                                   'Keyname', 'Keyword', 
                                   'Campaign', 'CreativeTarget', 
                                   'Creative', 'Notification'),
                    id, campaign_id=NULL)
}
\arguments{
\item{credentials}{a character string as returned by \link{oas_build_credentials}}

\item{request_type}{a character string in one of the supported 
object types for the API database read action}

\item{id}{a character string that uniquely identifies an item to retrieve. For the request
of  some objects Id serves as the Name (Affiliate, ConversionProcess, Event, Transaction, Keyname)}

\item{campaign_id}{a character string that identifies a campaign when CreativeTarget, Creative
or Notification reads are being performed, otherwise this value will be ignored}
}
\value{
A \code{list} of all fields available for the specified request type
}
\description{
This function returns a lists of all the available data points for 
requested data type specified in the read_request. Most elements support the 
READ function except Companion Position, Competitive Category and Code.
}
\examples{
\dontrun{
my_credentials <- build_credentials('myaccount', 
                                    'myusername', 
                                    'mypassword')
site_details <- oas_read(credentials=my_credentials, 
                                 request_type='Site', 
                                 id='www.mysite.com')
campaign_details <- oas_read(credentials=my_credentials, 
                                 request_type='Campaign', 
                                 id='one_campaign_id')
creative_details <- oas_read(credentials=my_credentials, 
                                 request_type='Creative', 
                                 id='one_creative_id', 
                                 campaign_id='one_campaign_id')
}
}
\concept{
api read
}

