% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_live.R
\name{oas_run_live}
\alias{oas_run_live}
\title{Execute Run Live Campaigns Command}
\usage{
oas_run_live(credentials, 
                    action = c('TestLiveCampaigns', 
                               'RunLiveCampaigns', 'LastStatusRLC'), 
                    verbose = FALSE)
}
\arguments{
\item{credentials}{a character string as returned by \link{oas_build_credentials}}

\item{action}{a character string in one of the supported run live operations}

\item{verbose}{a boolean indicating whether messages should be printed while making the request}
}
\value{
A \code{list} of the resulting operation
}
\description{
This function allows you to run 3 different RLC operations
}
\examples{
\dontrun{
my_credentials <- oas_build_credentials('myaccount', 
                                        'myusername', 
                                        'mypassword')
                                        
tlc_result <- oas_run_live(credentials, 
                           action='TestLiveCampaigns')
                           
rlc_result <- oas_run_live(credentials, 
                           action='RunLiveCampaigns')
                           
last_rlc_result <- oas_run_live(credentials, 
                                action='LastStatusRLC')
                                            
}
}
\concept{
api run_live
}

