% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{oas_update}
\alias{oas_update}
\title{Update Items in API Database}
\usage{
oas_update(credentials, 
                    request_type=c('Advertiser', 'AdvertiserCategory', 'Affiliate',
                                   'Agency', 'CampaignGroup', 'CompanionPosition',
                                   'CompetitiveCategory', 'ConversionProcess', 
                                   'CreativeType',
                                   'Event', 'InsertionOrder', 'Notification',
                                   'Page', 'Product', 'RichMediaTemplate', 
                                   'SalesPerson', 'Section', 'Site',
                                   'SiteGroup', 'Transaction', 'Position',
                                   'Keyword', 'Keyname', 'Publisher', 
                                   'Campaign', 
                                   'CreativeTarget', 'Creative'),
                     update_data, verbose = FALSE)
}
\arguments{
\item{credentials}{a character string as returned by \link{oas_build_credentials}}

\item{request_type}{a character string in one of the supported 
object types for the API database list action}

\item{update_data}{an XML document specifying the data to be updated based on the request_type}

\item{verbose}{a boolean indicating whether messages should be printed while making the request}
}
\value{
A \code{logical} being TRUE if the update was successful
}
\description{
This function updates items in Xaxis for Publishers.
}
\note{
We recommend using the \link{oas_list} function to first understand
what attributes are available to be updated on each object type in Xaxis for Publishers
before attempting to update existing objects.
}
\examples{
\dontrun{
my_credentials <- oas_build_credentials('myaccount', 
                                        'myusername', 
                                        'mypassword')
                                    
adver_update <- oas_update(credentials=my_credentials, 
                           request_type='Advertiser', 
                           update_data=list(addChildren(
                                            newXMLNode('Advertiser'), 
                                               list(newXMLNode('Id', 'thisadvertiserid'), 
                                                    newXMLNode('ContactTitle', 'new Title')))))
                                            
campaign_update <- oas_update(credentials=my_credentials, 
                              request_type='Campaign', 
                              update_data=list(addChildren(
                                 newXMLNode('Overview'), 
                                 list(newXMLNode('Id', 'myExistingCampaignId'), 
                                 newXMLNode('Status', 'L'),
                                 addChildren(
                                 newXMLNode('CompetitiveCategories'),
                                 list(newXMLNode('CompetitiveCategoryId','Airlines'), 
                                      newXMLNode('CompetitiveCategoryId','Travel')))))))
                                            
}
}
\concept{
api update
}

