% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{oas_zone_inventory}
\alias{oas_zone_inventory}
\title{Retrieve Zone Inventory Reports}
\usage{
oas_zone_inventory(credentials, 
                              report_name=c('Statistics by Zone', 
                                            'Forecast by Zone', 
                                            'Detail Statistics',
                                            'Detail Forecast'), 
                              site_id,
                              max_row="100",
                              zone_name=NULL,
                              start_date=NULL,
                              end_date=NULL)
}
\arguments{
\item{credentials}{a character string as returned by \link{build_credentials}}

\item{report_name}{a character string one of the valid zone inventory reports}

\item{site_id}{a character string containing an identifying name for a Site}

\item{max_row}{a character integer limiting the number of rows returned. 
A character is recommended to avoid larger numbers being formatted
in scientific notation and not being properly interpreted by the API.}

\item{zone_name}{a character string of a particular zone name for reporting on.}

\item{start_date}{a character string for inquiring inventory detail report. 
This field is optional and must be in the yyyy-mm-dd format.}

\item{end_date}{a character string for inquiring the inventory detail report by 
given schedule. This field is optional and must be in the yyyy-mm-dd format.}
}
\value{
A \code{data.frame} of zone inventory data in the format of the specified
report_name
}
\description{
This function returns a data.frame of inventory stats 
and forecasts broken down by a specified zone
}
\note{
Zone inventory requests are only available if the Custom Zip Zone module is enabled on the account
}
\examples{
\dontrun{
my_credentials <- build_credentials('myaccount', 
                                    'myusername', 
                                    'mypassword')
                                    
# Note that forecast reports must have start and end dates >= Sys.Date()
otherwise they will return 0 row data.frames                                  
zone_forecast <- oas_zone_inventory(credentials=my_credentials, 
                                        report_name='Forecast by Zone',
                                        site_id='www.mysite.com',
                                        max_row='20000',
                                        zone_name='UNKNOWN',
                                        start_date='2015-12-01', 
                                        end_date='2015-12-31')
                                     
}
}
\concept{
api inventory zone report
}

