% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perform_request}
\alias{perform_request}
\title{POST AdXML Request to API}
\usage{
perform_request(xmlBody)
}
\arguments{
\item{xmlBody}{character string formatted to look like XML in order to make the request}
}
\value{
A \code{list}, with header and text gatherers
}
\description{
This function is an internal helper actually makes the POST request
}
\concept{
api request
}

