% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.R
\name{report_request}
\alias{report_request}
\title{Run Template Reports}
\usage{
report_request(credentials,
                      report_type,
                      report_name,
                      id=NULL,
                      start_date=NULL,
                      end_date=NULL,
                      threshold=NULL)
}
\arguments{
\item{credentials}{a character string as returned by \link{build_credentials}}

\item{report_type}{a character string naming the type of report being requested}

\item{report_name}{a character string naming the Report Table}

\item{id}{a character string that identifies one object to report on, for example
a campaign id if reporting on one campaign's revenue}

\item{start_date}{a character string representing a date formatted as YYYY-MM-DD of
when to start the reporting from}

\item{end_date}{a character string representing a date formatted as YYYY-MM-DD of
when to end the reporting on}

\item{threshold}{a mandatory integer when requesting Over/Under Campaign Delivery Health
Status reports. Ignored otherwise.}
}
\value{
A \code{data.frame} of all fields available for the specified report name and type
}
\description{
This function runs a precanned report that is also available via
the OAS UI. Only one report is available at a time. The report_type
and report_name provided will be checked against the data.frame
available_reports before submitting requests, so please consult
that data if unsure whether a particular report is supported.
}
\note{
Please refer to the package data.frame \link{available_reports}
and/or the API documentation to determine the correct type and report_name combinations

For Site reports the Site's Domain has to be used as the Id
element (e.g. www.mysite.com)
}
\examples{
\dontrun{
my_credentials <- build_credentials('myaccount',
                                    'myusername',
                                    'mypassword')
site_delivery_info <- report_request(credentials=my_credentials,
                                     report_type='Site Delivery',
                                     report_name='Executive Summary',
                                     id='www.mysite.com',
                                     start_date='2015-09-01',
                                     end_date='2015-09-30')
pos_delivery_info <- report_request(credentials=my_credentials,
                                    report_type='Campaign Delivery',
                                    report_name='Position Delivery Information',
                                    id='one_campaign_id')
overunder_campaigns <- report_request(credentials=my_credentials,
                                      report_type='Daily Health Status',
                                      report_name='Over Delivery Campaigns',
                                      threshold=20)
}
}
\concept{
api report
}

