% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{request_builder}
\alias{request_builder}
\title{Build AdXML Request for API}
\usage{
request_builder(credentials, adxml_request)
}
\arguments{
\item{credentials}{a character string as returned by \link{build_credentials}}

\item{adxml_request}{an XML document formatted as AdXML that forms the basis of the API request. The AdXML will vary depending on the request being made.}
}
\value{
A \code{character}, formatted as XML, that is ready for a POST to the API endpoint
}
\description{
This function is an internal helper that combines the credentials 
string with the AdXML body to create a properly formatted API request.
}
\concept{
api request
}

