% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventory_reports}
\alias{inventory_reports}
\title{Inventory Reports Table}
\format{a \code{data.frame} with 121 rows and 5 variables}
\usage{
data(inventory_reports)
}
\description{
This data.frame contains all of the inventory reports available via
the \link{oas_basic_inventory},  \link{oas_search_inventory}
\link{oas_geo_inventory}, and \link{oas_zone_inventory} functions. 
Simply provide the same report_type and report_name as seen in this table
to retrive via those functions
}
\details{
\itemize{
  \item inventory_type: the type of inventory function being called
  (e.g. basic, search, geo, or zone)
  \item keywords_type: a type attribute tagged to the keywords node
  if used in a report
  \item report_type: the type of object the report covers
  \item report_id: an OAS id needed internally to request the report table
  \item report_name: a human readable name of the report
}
}
\examples{
\dontrun{
data(inventory_reports)

head(inventory_reports)
#   inventory_type keywords_type   report_type           report_id           report_name
# 1          basic               configuration       Configuration         configuration
# 2          basic                    overview  AllCampaignsBooked  all campaigns booked
# 3          basic                    overview    AllSitesForecast    all sites forecast
# 4          basic                    overview AllSectionsForecast all sections forecast
# 5          basic                    overview     GlobalInventory      global inventory
# 6          basic      campaign      campaign        ListOverview         list overview
}
}
\keyword{datasets}

